Refactored Alert component, optimize and clean up!
